+++ workspaces/airbrake/plugins/airbrake/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-airbrake
 
+## 0.19.0
+
+### Minor Changes
+
+- 3e91bc8: Backstage version bump to v1.49.2
+
 ## 0.18.0
 
 ### Minor Changes