+++ workspaces/argocd/plugins/argocd-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-argocd-common
 
+## 1.15.0
+
+### Minor Changes
+
+- 265995b: Backstage version bump to v1.49.2
+
+## 1.14.1
+
+### Patch Changes
+
+- dc8e24e: Backstage version bump to v1.48.5
+
 ## 1.14.0
 
 ### Minor Changes