+++ workspaces/graphiql/plugins/graphiql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-graphiql
 
+## 0.14.0
+
+### Minor Changes
+
+- 0bd863f: Backstage version bump to v1.49.2
+
 ## 0.13.0
 
 ### Minor Changes