+++ workspaces/quay/plugins/quay-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-quay-common
 
+## 1.19.0
+
+### Minor Changes
+
+- cfd52e1: Backstage version bump to v1.49.2
+
+## 1.18.1
+
+### Patch Changes
+
+- 46068bd: Backstage version bump to v1.48.5
+
 ## 1.18.0
 
 ### Minor Changes